pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0,<69.3.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dj3fb"
dynamic = ["version"]
requires-python = ">= 3.10"
dependencies = [
    "django",
    "faker",
    "factory_boy",
]
authors = [
  {name = "three-kinds", email = "3179158552@qq.com"},
]
description = "dj3fb can generate DjangoModelFactory for your django models by using factory_boy."
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: Django",
]

[project.urls]
Homepage = "https://github.com/three-kinds/dj3fb"
Source = "https://github.com/three-kinds/dj3fb"

[tool.setuptools.dynamic]
version = {attr = "dj3fb.__version__"}

[tool.setuptools.packages.find]
include = ["dj3fb*"]

[tool.coverage.run]
source = ["dj3fb"]

[tool.coverage.report]
exclude_lines = [
    "raise NotImplementedError"
]

[tool.mypy]
packages = [
    "dj3fb",
    "tests"
]
disable_error_code = [
    "import-untyped",
    "method-assign"
]

[tool.ruff]
src = [
    "dj3fb",
    "tests"
]
line-length = 120
